Migrate mdbook_runs_preprocessors to BookTest
